requirements.txt
langchain>=0.3.20
langchain-openai>=0.2.14
langchain-core>=0.3.28
langgraph>=0.2.55
openai>=1.58.1
python-dotenv>=1.0.0
pydantic>=2.10.5
